+++ src/index.ts
@@ -1,24 +1,26 @@
- import * as axisInterfaces from "./axis/axisInterfaces";
- import * as axis from "./axis/axis";
- import * as axisScale from "./axis/axisScale";
- import * as axisStyle from "./axis/axisStyle";
+import * as axisInterfaces from "./axis/axisInterfaces";
+import * as axis from "./axis/axis";
+import * as axisScale from "./axis/axisScale";
+import * as axisStyle from "./axis/axisStyle";
 
- import * as dataLabelArrangeGrid from "./dataLabel/dataLabelArrangeGrid";
- import * as dataLabelInterfaces from "./dataLabel/dataLabelInterfaces";
- import * as dataLabelManager from "./dataLabel/dataLabelManager";
- import * as dataLabelUtils from "./dataLabel/dataLabelUtils";
- import * as locationConverter from "./dataLabel/locationConverter";
+import * as dataLabelArrangeGrid from "./dataLabel/dataLabelArrangeGrid";
+import * as dataLabelInterfaces from "./dataLabel/dataLabelInterfaces";
+import * as dataLabelManager from "./dataLabel/dataLabelManager";
+import * as dataLabelUtils from "./dataLabel/dataLabelUtils";
+import * as locationConverter from "./dataLabel/locationConverter";
 
- import * as interactiveLegend from "./legend/interactiveLegend";
- import * as legend from "./legend/legend";
- import * as legendBehavior from "./legend/behavior/legendBehavior";
- import OpacityLegendBehavior from "./legend/behavior/opacityLegendBehavior";
- import * as legendData from "./legend/legendData";
- import * as legendInterfaces from "./legend/legendInterfaces";
- import * as legendPosition from "./legend/legendPosition";
- import * as svgLegend from "./legend/svgLegend";
+import * as interactiveLegend from "./legend/interactiveLegend";
+import * as legend from "./legend/legend";
+import * as legendBehavior from "./legend/behavior/legendBehavior";
+import OpacityLegendBehavior from "./legend/behavior/opacityLegendBehavior";
+import * as legendData from "./legend/legendData";
+import * as legendInterfaces from "./legend/legendInterfaces";
+import * as legendPosition from "./legend/legendPosition";
+import * as svgLegend from "./legend/svgLegend";
 
- export {
+import * as label from "./label/index";
+
+export {
     axisInterfaces,
     axis,
     axisScale,
@@ -29,6 +31,7 @@
     dataLabelUtils,
     locationConverter,
     interactiveLegend,
+    label,
     legend,
     legendBehavior,
     OpacityLegendBehavior,
@@ -36,4 +39,4 @@
     legendInterfaces,
     legendPosition,
     svgLegend
- };
+};