+++ src/label/index.ts
@@ -1,0 +1,15 @@
+import * as familyInfo from "./familyInfo";
+import * as fontProperties from "./fontProperties";
+import * as labelLayout from "./labelLayout";
+import * as labelUtils from "./labelUtils";
+import * as newDataLabelUtils from "./newDataLabelUtils";
+import * as units from "./units";
+
+export {
+    familyInfo,
+    fontProperties,
+    labelLayout,
+    labelUtils,
+    newDataLabelUtils,
+    units
+};