+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.2.0
+* Implements legend marker shapes
+* New Label Utils
+
 ## 2.1.0
 * Update packages to fix vulnerabilities